[fix] 서비스 리뷰  - 별점 데이터 수정( 5.0 -> 5 )  - 세차 리뷰 임시진입점 살림 (주석 상태)

+++ app/src/main/java/com/genesis/apps/ui/main/service/ServiceReviewActivity.java
@@ -162,7 +162,7 @@ public class ServiceReviewActivity extends SubActivity<ActivityServiceReviewBind
     private void onClickOkBtn() {
         Log.d(TAG, "onClickOkBtn: " + reviewType);
         //사용자가 입력한 별점, 기타의견 가져오기
-        String starRating = String.valueOf(ui.rbServiceReviewRatingbar.getRating());
+        String starRating = String.valueOf((int) ui.rbServiceReviewRatingbar.getRating());
         String reviewInput = ui.etServiceReviewInput.getText().toString();
 
         //입력이 너무 길면 자르...는 걸로 일단 코드에서 처리
